pyproject.toml
[tool.poetry]
name = "djangocms-rest"
version = "0.1.0"  # Will be read from djangocms_rest/__init__.py
description = "API endpoints for django CMS"
authors = ["Fabian Braun <fsbraun@gmx.de>"]
maintainers = ["Django CMS Association and contributors <info@django-cms.org>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/fsbraun/djangocms-rest"
repository = "https://github.com/fsbraun/djangocms-rest"
packages = [{include = "djangocms_rest"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django CMS",
    "Framework :: Django CMS :: 3.11",
    "Framework :: Django CMS :: 4.0",
    "Framework :: Django CMS :: 4.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
]

[tool.poetry.dependencies]
python = ">=3.9"
django-cms = ">=4.1.1"
djangorestframework = "*"
djangocms-link = "^5.0.0"
djangocms-text = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"
mypy = "^1.3.0"
django-stubs = "^4.2.3"
djangorestframework-stubs = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
Homepage = "https://github.com/fsbraun/djangocms-rest"
